zzz (#12)
Note: This was edited, original commit message below the line.
Instead of adding zzz_exploit.py, we changed it to adapt to
the existing base since in our repository it never got deleted
in the first place.

 * Updated mysmb.py module with more options
 * Improve zzz_exploit.py to use said options
 * Deleted send_and_execute.py since the functionality got (mostly)
   merged into zzz_exploit.py
 * Renamed eternalblue_checker.py to eternal_checker.py
 * Modified said checker to use mysmb.py module
 * Deleted binary mysmb.pyc

Beware: The original change added the options to use a custom port in
zzz_exploit.py but this was never actually wired up with mysmb.py:

	def exploit(target, port, ...):
		conn = MYSMB(target, port)

Compare with mysmb.py:

	class MYSMB(smb.SMB):
		def __init__(self, remote_host, use_ntlmv2=True, ...):
			...

The port will thus always be interpreted as the variable `use_ntlmv2`
and can never actually be set in any custom way 0:-)

This is now fixed, since we base the changes to mysmb.py on the ones
done by helvio.

-------

* Create LICENSE

* added initial exploit code

* Update README.md

* Update README.md

* added link to tutorial video

* Minor change for consistency w/ common tools (#1)

* Update listener_prep.sh

* Update shell_prep.sh

* Updated README to match recent consistency changes

* Update README.md

* add in groom connection arg to README for clarity and added TODO

* added option to select staged or stageless payload in shell_prep.sh (#8)

* added fixes to mysmb.py as reccommended by issue #4 (#9)

* Update README.md

* eternal blue exploit for windows 10, same code as 8 just with notes for my personal organizaton

* eternal blue exploit for windows 10, same code as 8 just with notes for my personal organizaton

* listener_prep.sh modified for stageless/staged payloads

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* added improved zzz_exploit.py and mysmb with more named pipes

* changed checker to use mysmb function

* Delete mysmb.pyc